Moved CubeBackgrounModel class to a more generic Cube class.
